Refactored code from ConsentBanner to lib/utilities
